move plugin reload command
