+++ src/main/java/com/voltskiya/misc/admin/PluginAdmin.java
@@ -4,6 +4,7 @@ import com.voltskiya.lib.AbstractModule;
 import com.voltskiya.misc.admin.console.command.CommandIntercepter;
 import com.voltskiya.misc.admin.console.command.CommandListenerCommand;
 import com.voltskiya.misc.admin.console.monitor.CommandTaskManager;
+import com.voltskiya.misc.admin.reload.CommandReload;
 import org.bukkit.event.HandlerList;
 
 public class PluginAdmin extends AbstractModule {
